+++ lib/kubernetes/api.ts
@@ -17,11 +17,10 @@
 import { logger } from "@atomist/automation-client";
 import * as k8s from "@kubernetes/client-node";
 import * as http from "http";
-import * as stringify from "json-stringify-safe";
 import * as request from "request";
-import { DeepPartial } from "ts-essentials";
 import { requestError } from "../support/error";
 import { defaultNamespace } from "./namespace";
+import { logObject } from "./resource";
 
 /** Response from methods that operate on an resource. */
 export interface K8sObjectResponse {
@@ -41,8 +40,34 @@ export interface K8sDeleteResponse {
     response: http.IncomingMessage;
 }
 
-// avoid https://github.com/kubernetes-client/javascript/issues/52
-export type K8sObject = DeepPartial<k8s.KubernetesObject>;
+/** Response from list API method. */
+export interface K8sApiResponse {
+    body: k8s.V1APIResourceList;
+    response: http.IncomingMessage;
+}
+
+/** Union type of all response types. */
+type K8sRequestResponse = K8sObjectResponse | K8sDeleteResponse | K8sListResponse | K8sApiResponse;
+
+/** Kubernetes API verbs. */
+export type K8sApiAction = "create" | "delete" | "list" | "patch" | "read" | "replace";
+
+/** Type of option argument for object API requests. */
+export interface K8sObjectRequestOptions { headers: { [name: string]: string; }; }
+
+/**
+ * Valid Content-Type header values for patch operations.  See
+ * https://kubernetes.io/docs/tasks/run-application/update-api-object-kubectl-patch/
+ * for details.
+ */
+export enum K8sPatchStrategies {
+    /** Diff-like JSON format. */
+    JsonPatch = "application/json-patch+json",
+    /** Simple merge. */
+    MergePatch = "application/merge-patch+json",
+    /** Merge with different strategies depending on field metadata. */
+    StrategicMergePatch = "application/strategic-merge-patch+json",
+}
 
 /**
  * Dynamically construct Kubernetes API request URIs so client does
@@ -51,75 +76,94 @@ export type K8sObject = DeepPartial<k8s.KubernetesObject>;
  */
 export class K8sObjectApi extends k8s.ApisApi {
 
-    private readonly defaultDeleteBody: any = { propagationPolicy: "Background" };
+    private static readonly defaultDeleteBody: k8s.V1DeleteOptions = { propagationPolicy: "Background" };
 
     /**
      * Read any Kubernetes resource.
      */
-    public async create(spec: K8sObject): Promise<K8sObjectResponse> {
-        const requestOptions = this.baseRequestOptions("POST");
-        requestOptions.uri += specUriPath(spec, "create");
+    public async create(spec: k8s.KubernetesObject, options?: K8sObjectRequestOptions): Promise<K8sObjectResponse> {
+        const requestOptions = this.baseRequestOptions("POST", options);
+        requestOptions.uri += await this.specUriPath(spec, "create");
         requestOptions.body = spec;
-        return this.requestPromise(requestOptions) as any as K8sObjectResponse;
+        return this.requestPromise(requestOptions) as unknown as K8sObjectResponse;
     }
 
     /**
      * Delete any Kubernetes resource.
      */
-    public async delete(spec: K8sObject, body: any = this.defaultDeleteBody): Promise<K8sDeleteResponse> {
-        const requestOptions = this.baseRequestOptions("DELETE");
-        requestOptions.uri += specUriPath(spec, "delete");
-        requestOptions.body = body;
-        return this.requestPromise(requestOptions) as any as K8sDeleteResponse;
+    public async delete(spec: k8s.KubernetesObject, body?: k8s.V1DeleteOptions, options?: K8sObjectRequestOptions): Promise<K8sDeleteResponse> {
+        const requestOptions = this.baseRequestOptions("DELETE", options);
+        requestOptions.uri += await this.specUriPath(spec, "delete");
+        requestOptions.body = body || K8sObjectApi.defaultDeleteBody;
+        return this.requestPromise(requestOptions) as unknown as K8sDeleteResponse;
     }
 
     /**
      * List any Kubernetes resource.
      */
-    public async list(spec: K8sObject): Promise<K8sListResponse> {
-        const requestOptions = this.baseRequestOptions();
-        requestOptions.uri += specUriPath(spec, "list");
-        return this.requestPromise(requestOptions) as any as K8sListResponse;
+    public async list(spec: k8s.KubernetesObject, options?: K8sObjectRequestOptions): Promise<K8sListResponse> {
+        const requestOptions = this.baseRequestOptions("GET", options);
+        requestOptions.uri += await this.specUriPath(spec, "list");
+        return this.requestPromise(requestOptions) as unknown as K8sListResponse;
     }
 
     /**
      * Patch any Kubernetes resource.
      */
-    public async patch(spec: K8sObject): Promise<K8sObjectResponse> {
-        const requestOptions = this.baseRequestOptions("PATCH");
-        requestOptions.uri += specUriPath(spec, "patch");
+    public async patch(spec: k8s.KubernetesObject, options?: K8sObjectRequestOptions): Promise<K8sObjectResponse> {
+        const requestOptions = this.baseRequestOptions("PATCH", options);
+        requestOptions.uri += await this.specUriPath(spec, "patch");
         requestOptions.body = spec;
-        requestOptions.headers["Content-Type"] = "application/strategic-merge-patch+json";
-        return this.requestPromise(requestOptions) as any as K8sObjectResponse;
+        return this.requestPromise(requestOptions) as unknown as K8sObjectResponse;
     }
 
     /**
      * Read any Kubernetes resource.
      */
-    public async read(spec: K8sObject): Promise<K8sObjectResponse> {
-        const requestOptions = this.baseRequestOptions();
-        requestOptions.uri += specUriPath(spec, "read");
-        return this.requestPromise(requestOptions) as any as K8sObjectResponse;
+    public async read(spec: k8s.KubernetesObject, options?: K8sObjectRequestOptions): Promise<K8sObjectResponse> {
+        const requestOptions = this.baseRequestOptions("GET", options);
+        requestOptions.uri += await this.specUriPath(spec, "read");
+        return this.requestPromise(requestOptions) as unknown as K8sObjectResponse;
     }
 
     /**
      * Replace any Kubernetes resource.
      */
-    public async replace(spec: K8sObject): Promise<K8sObjectResponse> {
-        const requestOptions = this.baseRequestOptions("PUT");
-        requestOptions.uri += specUriPath(spec, "replace");
+    public async replace(spec: k8s.KubernetesObject, options?: K8sObjectRequestOptions): Promise<K8sObjectResponse> {
+        const requestOptions = this.baseRequestOptions("PUT", options);
+        requestOptions.uri += await this.specUriPath(spec, "replace");
         requestOptions.body = spec;
-        return this.requestPromise(requestOptions) as any as K8sObjectResponse;
+        return this.requestPromise(requestOptions) as unknown as K8sObjectResponse;
     }
 
     /**
-     * Generate request options.  Largely copied from @kubernetes/client-node/dist/api.js.
+     * Get metadata from Kubernetes API for resources described by
+     * `kind` and `apiVersion`.  If it is unable to find the resource
+     * `kind` under the provided `apiVersion` or an error occurs,
+     * `undefined` is returned.
      */
-    private baseRequestOptions(method: string = "GET"): request.UriOptions & request.CoreOptions {
+    public async resource(apiVersion: string, kind: string): Promise<k8s.V1APIResource | undefined> {
+        try {
+            const requestOptions = this.baseRequestOptions();
+            const prefix = (apiVersion.includes("/")) ? "apis" : "api";
+            requestOptions.uri += [prefix, apiVersion].join("/");
+            const getApiResponse = await this.requestPromise(requestOptions);
+            const apiResourceList = getApiResponse.body as unknown as k8s.V1APIResourceList;
+            return apiResourceList.resources.find(r => r.kind === kind);
+        } catch (e) {
+            logger.error(`Failed to fetch resource metadata for ${apiVersion}/${kind}: ${e.message}`);
+            return undefined;
+        }
+    }
+
+    /**
+     * Generate request options.  Largely copied from the common
+     * elements of @kubernetes/client-node action methods.
+     */
+    public baseRequestOptions(method: string = "GET", options?: K8sObjectRequestOptions): request.UriOptions & request.CoreOptions {
         const localVarPath = this.basePath + "/";
         const queryParameters = {};
-        const localHeaders = (method === "PATCH") ? { "Content-Type": "application/strategic-merge-patch+json" } : {};
-        const headerParams = Object.assign({}, this.defaultHeaders, localHeaders);
+        const headerParams = Object.assign({}, this.defaultHeaders, K8sObjectApi.methodHeaders(method), options?.headers || {});
         const requestOptions = {
             method,
             qs: queryParameters,
@@ -134,9 +178,52 @@ export class K8sObjectApi extends k8s.ApisApi {
     }
 
     /**
+     * Use spec information to construct resource URI path.  If any
+     * required information in not provided, an Error is thrown.  If an
+     * `apiVersion` is not provided, "v1" is used.  If a `metadata.namespace`
+     * is not provided for a request that requires one, "default" is used.
+     *
+     * @param spec resource spec which must kind and apiVersion properties
+     * @param action API action, see [[K8sApiAction]]
+     * @return tail of resource-specific URI
+     */
+    public async specUriPath(spec: k8s.KubernetesObject, action: K8sApiAction): Promise<string> {
+        if (!spec.kind) {
+            throw new Error(`Spec does not contain kind: ${logObject(spec)}`);
+        }
+        if (!spec.apiVersion) {
+            spec.apiVersion = "v1";
+            logger.info(`Spec does not contain apiVersion, using "${spec.apiVersion}"`);
+        }
+        if (!spec.metadata) {
+            spec.metadata = {};
+        }
+        const resource = await this.resource(spec.apiVersion, spec.kind);
+        if (!resource) {
+            throw new Error(`Unrecognized API version and kind: ${spec.apiVersion} ${spec.kind}`);
+        }
+        if (namespaceRequired(resource, action) && !spec.metadata.namespace) {
+            spec.metadata.namespace = defaultNamespace;
+        }
+        const prefix = (spec.apiVersion.includes("/")) ? "apis" : "api";
+        const parts = [prefix, spec.apiVersion];
+        if (resource.namespaced && spec.metadata.namespace) {
+            parts.push("namespaces", spec.metadata.namespace);
+        }
+        parts.push(resource.name);
+        if (appendName(action)) {
+            if (!spec.metadata.name) {
+                throw new Error(`Spec does not contain name: ${logObject(spec)}`);
+            }
+            parts.push(spec.metadata.name);
+        }
+        return parts.join("/").toLowerCase();
+    }
+
+    /**
      * Wrap request in a Promise.  Largely copied from @kubernetes/client-node/dist/api.js.
      */
-    private requestPromise(requestOptions: request.UriOptions & request.CoreOptions): Promise<K8sObjectResponse | K8sDeleteResponse> {
+    private requestPromise(requestOptions: request.UriOptions & request.CoreOptions): Promise<K8sRequestResponse> {
         return new Promise((resolve, reject) => {
             request(requestOptions, (error, response, body) => {
                 if (error) {
@@ -152,113 +239,38 @@ export class K8sObjectApi extends k8s.ApisApi {
         });
     }
 
-}
-
-/** Options for creating a Kubernetes resource API path from a spec. */
-export interface SpecUriPathOptions {
-    /** If true, append resource name to URL path. */
-    appendName?: boolean;
-    /** If true, ensure path contains a namespace. */
-    namespaceRequired?: boolean;
-}
-
-export type K8sApiAction = "create" | "delete" | "list" | "patch" | "read" | "replace";
-
-/**
- * Generate proper [[SpecUriPathOptions]] from API action and resource kind.
- *
- * @param action API HTTP action
- * @param kind Kubernetes resource kind
- * @return Kubernetes API URI path generation options
- */
-export function uriOpts(action: K8sApiAction, kind: string): SpecUriPathOptions {
-    const appendName = (action === "create" || action === "list") ? false : true;
-    const namespaceRequired = !(action === "list" || isClusterResource(action, kind));
-    return { appendName, namespaceRequired };
-}
-
-export function isClusterResource(action: K8sApiAction, kind: string): boolean {
-    const clusterResources = [
-        "APIService",
-        "AuditSink",
-        "CertificateSigningRequest",
-        "ClusterCustomObject",
-        "ClusterRole",
-        "ClusterRoleBinding",
-        "CustomResourceDefinition",
-        "InitializerConfiguration",
-        "MutatingWebhookConfiguration",
-        "Namespace",
-        "Node",
-        "PersistentVolume",
-        "PodSecurityPolicy",
-        "PriorityClass",
-        "SelfSubjectAccessReview",
-        "SelfSubjectRulesReview",
-        "StorageClass",
-        "SubjectAccessReview",
-        "TokenReview",
-        "ValidatingWebhookConfiguration",
-        "VolumeAttachment",
-    ];
-    const clusterStatuses = [
-        "APIServiceStatus",
-        "CertificateSigningRequestStatus",
-        "CustomResourceDefinitionStatus",
-        "NamespaceStatus",
-        "NodeStatus",
-        "PersistentVolumeStatus",
-        "VolumeAttachmentStatus",
-    ];
-    if (clusterResources.includes(kind)) {
-        return true;
-    } else if (action === "patch" || action === "replace") {
-        return clusterStatuses.includes(kind);
-    } else if (action === "read") {
-        return [...clusterStatuses, "ComponentStatus"].includes(kind);
-    } else if (action === "list" && kind === "ComponentStatus") {
-        return true;
+    /**
+     * Return default headers based on action.
+     */
+    private static methodHeaders(method: string): { [name: string]: string } {
+        return (method === "PATCH") ? { "Content-Type": K8sPatchStrategies.StrategicMergePatch } : {};
     }
-    return false;
 }
 
 /**
- * Use spec information to construct resource URI path.  If any
- * required information in not provided, an Error is thrown.  If an
- * `apiVersion` is not provided, "v1" is used.  If a `metadata.namespace`
- * is not provided for a request that requires one, "default" is used.
+ * Return whether the name of the resource should be appended to the
+ * API URI path.  When creating and listing resources, it is not
+ * appended.
  *
- * @param spec resource metadata
- * @param appendName if `true`, append name to path
- * @return tail of resource-specific URI
+ * @param action API action, see [[K8sApiAction]]
+ * @return true if name should be appended to URI
  */
-export function specUriPath(spec: K8sObject, action: K8sApiAction): string {
-    if (!spec.kind) {
-        throw new Error(`Spec does not contain kind: ${stringify(spec)}`);
-    }
-    if (!spec.apiVersion) {
-        spec.apiVersion = "v1";
-        logger.info(`Spec does not contain apiVersion, using "${spec.apiVersion}"`);
-    }
-    if (!spec.metadata) {
-        spec.metadata = {};
-    }
-    const opts = uriOpts(action, spec.kind);
-    if (opts.namespaceRequired && !spec.metadata.namespace) {
-        spec.metadata.namespace = defaultNamespace;
-    }
-    const plural = spec.kind.toLowerCase().replace(/s$/, "se").replace(/y$/, "ie") + "s";
-    const prefix = (spec.apiVersion.includes("/")) ? "apis" : "api";
-    const parts = [prefix, spec.apiVersion];
-    if (spec.metadata.namespace) {
-        parts.push("namespaces", spec.metadata.namespace);
-    }
-    parts.push(plural);
-    if (opts.appendName) {
-        if (!spec.metadata.name) {
-            throw new Error(`Spec does not contain name: ${stringify(spec)}`);
-        }
-        parts.push(spec.metadata.name);
-    }
-    return parts.join("/").toLowerCase();
+export function appendName(action: K8sApiAction): boolean {
+    return !(action === "create" || action === "list");
+}
+
+/**
+ * Return whether namespace must be included in resource API URI.
+ * It returns true of the resource is namespaced and the action is
+ * not "list".  The namespace can be provided when the action is
+ * "list", but it need not be.
+ *
+ * @param resource resource metadata
+ * @param action API action, see [[K8sApiAction]]
+ * @return true is the namespace is required in the API URI path
+ */
+export function namespaceRequired(resource: k8s.V1APIResource, action: K8sApiAction): boolean {
+    // return action !== "list" || resource.namespaced;
+    // return !(action === "list" || !resource.namespaced);
+    return resource.namespaced && action !== "list";
 }