+++ lib/kubernetes/roleBinding.ts
@@ -17,18 +17,18 @@
 import { logger } from "@atomist/automation-client";
 import * as k8s from "@kubernetes/client-node";
 import * as _ from "lodash";
-import { DeepPartial } from "ts-essentials";
 import { errMsg } from "../support/error";
 import { logRetry } from "../support/retry";
 import { applicationLabels } from "./labels";
 import { metadataTemplate } from "./metadata";
+import { patchHeaders } from "./patch";
 import {
     appName,
     KubernetesApplication,
     KubernetesResourceRequest,
     KubernetesSdm,
 } from "./request";
-import { stringifyObject } from "./resource";
+import { logObject } from "./resource";
 
 /**
  * Create or patch role or cluster rolebinding.
@@ -44,14 +44,14 @@ export async function upsertRoleBinding(req: KubernetesResourceRequest): Promise
             await req.clients.rbac.readClusterRoleBinding(spec.metadata.name);
         } catch (e) {
             logger.debug(`Failed to read cluster role binding ${slug}, creating: ${errMsg(e)}`);
-            logger.info(`Creating cluster role binding ${slug} using '${stringifyObject(spec)}'`);
+            logger.info(`Creating cluster role binding ${slug} using '${logObject(spec)}'`);
             await logRetry(() => req.clients.rbac.createClusterRoleBinding(spec),
                 `create cluster role binding ${slug}`);
             return spec;
         }
-        logger.info(`Cluster role binding ${slug} exists, patching using '${stringifyObject(spec)}'`);
-        await logRetry(() => req.clients.rbac.patchClusterRoleBinding(spec.metadata.name, spec),
-            `patch cluster role binding ${slug}`);
+        logger.info(`Cluster role binding ${slug} exists, patching using '${logObject(spec)}'`);
+        await logRetry(() => req.clients.rbac.patchClusterRoleBinding(spec.metadata.name, spec,
+            undefined, undefined, undefined, undefined, patchHeaders()), `patch cluster role binding ${slug}`);
         return spec;
     } else {
         const spec = await roleBindingTemplate(req);
@@ -59,14 +59,14 @@ export async function upsertRoleBinding(req: KubernetesResourceRequest): Promise
             await req.clients.rbac.readNamespacedRoleBinding(spec.metadata.name, spec.metadata.namespace);
         } catch (e) {
             logger.debug(`Failed to read role binding ${slug}, creating: ${errMsg(e)}`);
-            logger.info(`Creating role binding ${slug} using '${stringifyObject(spec)}'`);
+            logger.info(`Creating role binding ${slug} using '${logObject(spec)}'`);
             await logRetry(() => req.clients.rbac.createNamespacedRoleBinding(spec.metadata.namespace, spec),
                 `create role binding ${slug}`);
             return spec;
         }
-        logger.info(`Role binding ${slug} exists, patching using '${stringifyObject(spec)}'`);
-        await logRetry(() => req.clients.rbac.patchNamespacedRoleBinding(spec.metadata.name, spec.metadata.namespace, spec),
-            `patch role binding ${slug}`);
+        logger.info(`Role binding ${slug} exists, patching using '${logObject(spec)}'`);
+        await logRetry(() => req.clients.rbac.patchNamespacedRoleBinding(spec.metadata.name, spec.metadata.namespace, spec,
+            undefined, undefined, undefined, undefined, patchHeaders()), `patch role binding ${slug}`);
         return spec;
     }
 }
@@ -93,8 +93,7 @@ export async function roleBindingTemplate(req: KubernetesApplication & Kubernete
     });
     const apiVersion = "rbac.authorization.k8s.io/v1";
     const kind = "RoleBinding";
-    // avoid https://github.com/kubernetes-client/javascript/issues/52
-    const rb: DeepPartial<k8s.V1RoleBinding> = {
+    const rb: k8s.V1RoleBinding = {
         apiVersion,
         kind,
         metadata,
@@ -117,7 +116,7 @@ export async function roleBindingTemplate(req: KubernetesApplication & Kubernete
         _.merge(rb, req.roleBindingSpec, { apiVersion, kind });
         rb.metadata.namespace = req.ns;
     }
-    return rb as k8s.V1RoleBinding;
+    return rb;
 }
 
 /**
@@ -141,8 +140,7 @@ export async function clusterRoleBindingTemplate(req: KubernetesApplication & Ku
     });
     const apiVersion = "rbac.authorization.k8s.io/v1";
     const kind = "ClusterRoleBinding";
-    // avoid https://github.com/kubernetes-client/javascript/issues/52
-    const rb: DeepPartial<k8s.V1ClusterRoleBinding> = {
+    const rb: k8s.V1ClusterRoleBinding = {
         apiVersion,
         kind,
         metadata,
@@ -165,5 +163,5 @@ export async function clusterRoleBindingTemplate(req: KubernetesApplication & Ku
     if (req.roleBindingSpec) {
         _.merge(rb, req.roleBindingSpec, { apiVersion, kind });
     }
-    return rb as k8s.V1ClusterRoleBinding;
+    return rb;
 }