+++ lib/kubernetes/service.ts
@@ -17,7 +17,6 @@
 import { logger } from "@atomist/automation-client";
 import * as k8s from "@kubernetes/client-node";
 import * as _ from "lodash";
-import { DeepPartial } from "ts-essentials";
 import { errMsg } from "../support/error";
 import { logRetry } from "../support/retry";
 import {
@@ -25,13 +24,14 @@ import {
     matchLabels,
 } from "./labels";
 import { metadataTemplate } from "./metadata";
+import { patchHeaders } from "./patch";
 import {
     appName,
     KubernetesApplication,
     KubernetesResourceRequest,
     KubernetesSdm,
 } from "./request";
-import { stringifyObject } from "./resource";
+import { logObject } from "./resource";
 
 /**
  * If `req.port` is truthy, create a service if it does not exist and
@@ -52,13 +52,13 @@ export async function upsertService(req: KubernetesResourceRequest): Promise<k8s
         await req.clients.core.readNamespacedService(spec.metadata.name, spec.metadata.namespace);
     } catch (e) {
         logger.debug(`Failed to read service ${slug}, creating: ${errMsg(e)}`);
-        logger.info(`Creating service ${slug} using '${stringifyObject(spec)}'`);
+        logger.info(`Creating service ${slug} using '${logObject(spec)}'`);
         await logRetry(() => req.clients.core.createNamespacedService(spec.metadata.namespace, spec), `create service ${slug}`);
         return spec;
     }
-    logger.info(`Service ${slug} exists, patching using '${stringifyObject(spec)}'`);
-    await logRetry(() => req.clients.core.patchNamespacedService(spec.metadata.name, spec.metadata.namespace, spec),
-        `patch service ${slug}`);
+    logger.info(`Service ${slug} exists, patching using '${logObject(spec)}'`);
+    await logRetry(() => req.clients.core.patchNamespacedService(spec.metadata.name, spec.metadata.namespace, spec,
+        undefined, undefined, undefined, undefined, patchHeaders()), `patch service ${slug}`);
     return spec;
 }
 
@@ -84,8 +84,7 @@ export async function serviceTemplate(req: KubernetesApplication & KubernetesSdm
     });
     const apiVersion = "v1";
     const kind = "Service";
-    // avoid https://github.com/kubernetes-client/javascript/issues/52
-    const s: DeepPartial<k8s.V1Service> = {
+    const s: k8s.V1Service = {
         apiVersion,
         kind,
         metadata,
@@ -105,5 +104,5 @@ export async function serviceTemplate(req: KubernetesApplication & KubernetesSdm
         _.merge(s, req.serviceSpec, { apiVersion, kind });
         s.metadata.namespace = req.ns;
     }
-    return s as k8s.V1Service;
+    return s;
 }