+++ CHANGELOG.md
@@ -17,6 +17,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   Do not supply defaults in resource spec templates. [713120e](https://github.com/atomist/sdm-pack-k8s/commit/713120ece4163c9693c8a5c2fed322ae7162d78b)
 -   Use policy/v1beta1 API for PodSecurityPolicy. [afc06d7](https://github.com/atomist/sdm-pack-k8s/commit/afc06d77e4f40f8899cb9e2bdc34ec84230ae82d)
 -   Do not terminate app deployment if namespace patching fails. [#79](https://github.com/atomist/sdm-pack-k8s/issues/79)
+-   Shorten k8s object spec strings when logged. [52f16bb](https://github.com/atomist/sdm-pack-k8s/commit/52f16bb1722ec7771fe54cae66005497bb0d2428)
+-   Remove patch header patch. [bd9aedd](https://github.com/atomist/sdm-pack-k8s/commit/bd9aedd5865607fc95ee5d839718b721a90ccfcc)
+-   Add environment to display name. [949b583](https://github.com/atomist/sdm-pack-k8s/commit/949b5836c611f888e8664fc69e5b833354bd36e1)
+-   Add headers argument to K8sObjectApi methods. [2f20455](https://github.com/atomist/sdm-pack-k8s/commit/2f204555549264b1a64903ebff3e059c2102f37a)
+-   **BREAKING** Change sync patch strategy to JSON merge patch. [04032e9](https://github.com/atomist/sdm-pack-k8s/commit/04032e9bd018ca088d6179637d5bb0741a2c5dc7)
+
+### Deprecated
+
+-   Remove unneeded uses of DeepPartial. [11b84dd](https://github.com/atomist/sdm-pack-k8s/commit/11b84dd6267db50d00cba4122af307da5c8ca28a)
+
+### Fixed
+
+-   Fix multi-document YAML support in sync. [14bd761](https://github.com/atomist/sdm-pack-k8s/commit/14bd761864414e1bc7a0c452d4719366a8704c64)
+-   Dynamic API does not support cluster-scoped custom resource definitions. [#82](https://github.com/atomist/sdm-pack-k8s/issues/82)
 
 ## [1.10.2](https://github.com/atomist/sdm-pack-k8s/compare/1.10.1...1.10.2) - 2019-09-27